Split sql2 logical plan ownership from datafusion
